+++ lib/src/locale_switch_sub_widgets/select_locale_button.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import 'package:locale_switcher/locale_switcher.dart';
+import 'package:locale_switcher/src/locale_store.dart';
+
+/// IconButton to show and select a language.
+///
+/// In popup window will be displayed [LocaleSwitcher.grid].
+class SelectLocaleButton extends StatelessWidget {
+  final bool updateIconOnChange;
+
+  final String toolTipPrefix;
+
+  final double radius;
+
+  final Widget? useStaticIcon;
+
+  final String popUpWindowTitle;
+
+  const SelectLocaleButton({
+    super.key,
+    this.updateIconOnChange = true,
+    this.toolTipPrefix = 'Current language: ',
+    this.radius = 32,
+    this.useStaticIcon,
+    this.popUpWindowTitle = "",
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ValueListenableBuilder(
+      valueListenable: LocaleStore.realLocaleNotifier,
+      builder: (BuildContext context, value, Widget? child) {
+        return IconButton(
+          icon: useStaticIcon ??
+              LangIconWithToolTip(
+                // langCode: LocaleManager.locale.value.languageCode,
+                toolTipPrefix: toolTipPrefix,
+                langCode: LocaleStore.realLocaleNotifier.value,
+                radius: radius,
+              ),
+          // tooltip: LocaleStore.languageToCountry[showOtherLocales]?[1] ??
+          //     "Other locales",
+          onPressed: () =>
+              showSelectLocaleDialog(context, title: popUpWindowTitle),
+        );
+      },
+    );
+  }
+}