+++ lib/src/locale_switcher.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:locale_switcher/src/locale_store.dart';
 import 'package:locale_switcher/src/locale_switch_sub_widgets/drop_down_menu_language_switch.dart';
 import 'package:locale_switcher/src/locale_switch_sub_widgets/grid_of_languages.dart';
+import 'package:locale_switcher/src/locale_switch_sub_widgets/select_locale_button.dart';
 import 'package:locale_switcher/src/locale_switch_sub_widgets/toggle_language_switch.dart';
 
 const showOtherLocales = 'show_other_locales';
@@ -12,6 +13,7 @@ enum _Switcher {
   menu,
   custom,
   grid,
+  iconButton,
 }
 
 typedef LocaleSwitchBuilder = Widget Function(List<String>);
@@ -67,10 +69,20 @@ class LocaleSwitcher extends StatelessWidget {
   /// ```
   final LocaleSwitchBuilder? builder;
 
-  /// Only for LocaleSwitcher.grid constructor
+  /// Only for [LocaleSwitcher.grid] constructor
   final SliverGridDelegate? gridDelegate;
 
+  /// Only for [LocaleSwitcher.grid] constructor
   final Function(BuildContext)? additionalCallBack;
+
+  /// Only for [LocaleSwitcher.iconButton] - use static icon.
+  final Icon? useStaticIcon;
+
+  /// Only for [LocaleSwitcher.iconButton].
+  final String? toolTipPrefix;
+
+  /// Only for [LocaleSwitcher.iconButton].
+  final double? iconRadius;
 
   /// A Widget to switch locale of App.
   const LocaleSwitcher._({
@@ -86,9 +98,14 @@ class LocaleSwitcher extends StatelessWidget {
     this.builder,
     this.gridDelegate,
     this.additionalCallBack,
+    this.toolTipPrefix,
+    this.useStaticIcon,
+    this.iconRadius,
   }) : _type = type;
 
-  /// A Widget to switch locale of App with [ToggleLanguageSwitch].
+  /// A Widget to switch locale of App with [AnimatedToggleSwitch](https://pub.dev/documentation/animated_toggle_switch/latest/animated_toggle_switch/AnimatedToggleSwitch-class.html).
+  ///
+  /// Example: [online app](https://alexqwesa.github.io/locale_switcher/), [source code](https://github.com/Alexqwesa/locale_switcher/blob/main/example/lib/main.dart).
   factory LocaleSwitcher.toggle({
     Key? key,
     String? title = 'Choose language:',
@@ -112,7 +129,9 @@ class LocaleSwitcher extends StatelessWidget {
     );
   }
 
-  /// A Widget to switch locale of App with [DropDownMenuLanguageSwitch]
+  /// A Widget to switch locale of App with [DropDownMenu](https://api.flutter.dev/flutter/material/DropdownMenu-class.html).
+  ///
+  /// Example: [online app](https://alexqwesa.github.io/locale_switcher/), [source code](https://github.com/Alexqwesa/locale_switcher/blob/main/example/lib/main.dart).
   factory LocaleSwitcher.menu({
     Key? key,
     String? title = 'Choose language:',
@@ -132,7 +151,10 @@ class LocaleSwitcher extends StatelessWidget {
     );
   }
 
-  /// A Widget to switch locale of App with [DropDownMenuLanguageSwitch]
+  /// A Widget to switch locale of App with [GridView](https://api.flutter.dev/flutter/widgets/GridView-class.html).
+  ///
+  /// Example: [online app](https://alexqwesa.github.io/locale_switcher/),
+  /// [source code](https://github.com/Alexqwesa/locale_switcher/blob/main/example/lib/main.dart) - click on icon in AppBar to see this widget.
   factory LocaleSwitcher.grid({
     Key? key,
     int numberOfShown = 200,
@@ -151,7 +173,6 @@ class LocaleSwitcher extends StatelessWidget {
   }
 
   /// A Widget to switch locale of App with your own widget:
-  ///
   ///
   /// Example:
   /// ```dart
@@ -187,11 +208,40 @@ class LocaleSwitcher extends StatelessWidget {
         builder: builder);
   }
 
+  /// A Widget to switch locale of App with [GridView](https://api.flutter.dev/flutter/widgets/GridView-class.html).
+  ///
+  /// Example: [online app](https://alexqwesa.github.io/locale_switcher/),
+  /// [source code](https://github.com/Alexqwesa/locale_switcher/blob/main/example/lib/main.dart) - it is an icon in AppBar.
+  ///
+  /// In popup window will be displayed [LocaleSwitcher.grid].
+  factory LocaleSwitcher.iconButton({
+    Key? key,
+    String? toolTipPrefix = 'Current language: ',
+    String? title = 'Select language: ',
+    Icon? useStaticIcon,
+    double? iconRadius = 32,
+    // required LocaleSwitchBuilder builder,
+    int numberOfShown = 200,
+    bool showOsLocale = true,
+  }) {
+    return LocaleSwitcher._(
+      key: key,
+      title: title,
+      toolTipPrefix: toolTipPrefix,
+      showOsLocale: showOsLocale,
+      numberOfShown: numberOfShown,
+      useStaticIcon: useStaticIcon,
+      iconRadius: iconRadius,
+      type: _Switcher.iconButton,
+      // builder: builder,
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-// todo: move to initState ?
+    // todo: move to initState ?
     if (LocaleStore.supportedLocales.isEmpty) {
-// assume it was not inited
+      // assume it was not inited
       final child = context.findAncestorWidgetOfExactType<MaterialApp>() ??
           context.findAncestorWidgetOfExactType<CupertinoApp>();
       if (child != null) {
@@ -215,16 +265,9 @@ class LocaleSwitcher extends StatelessWidget {
       }
     }
 
-// Prepare list of languageCodes where systemLocale is first and length == numberOfShown
-// final systemLocaleExist = (LocaleStore.supportedLocales ?? []).where(
-//   (element) => element == WidgetsBinding.instance.platformDispatcher.locale,
-// );
     final staticLocales = <String>[
       if (showOsLocale) LocaleStore.systemLocale,
-// if (systemLocaleExist.isNotEmpty) systemLocaleExist.first.languageCode,
       ...LocaleStore.supportedLocales
-// .where((element) =>
-//     element != WidgetsBinding.instance.platformDispatcher.locale)
           .take(numberOfShown) // chose most used
           .map((e) => e.languageCode),
     ];
@@ -241,6 +284,13 @@ class LocaleSwitcher extends StatelessWidget {
         }
 
         return switch (_type) {
+          _Switcher.custom => builder!(locales),
+          _Switcher.menu =>
+            DropDownMenuLanguageSwitch(locales: locales, title: title),
+          _Switcher.grid => GridOfLanguages(
+              gridDelegate: gridDelegate,
+              additionalCallBack: additionalCallBack,
+            ),
           _Switcher.toggle => ToggleLanguageSwitch(
               padding: padding,
               crossAxisAlignment: crossAxisAlignment,
@@ -249,14 +299,12 @@ class LocaleSwitcher extends StatelessWidget {
               title: title,
               locales: locales,
             ),
-          _Switcher.menu =>
-            DropDownMenuLanguageSwitch(locales: locales, title: title),
-          _Switcher.custom => builder!(locales),
-          // TODO: Handle this case.
-          _Switcher.grid => GridOfLanguages(
-              gridDelegate: gridDelegate,
-              additionalCallBack: additionalCallBack,
-            ),
+          _Switcher.iconButton => SelectLocaleButton(
+              radius: iconRadius ?? 32,
+              popUpWindowTitle: title ?? '',
+              updateIconOnChange: (useStaticIcon != null),
+              useStaticIcon: useStaticIcon,
+              toolTipPrefix: toolTipPrefix ?? ''),
         };
       },
     );